// frontend/src/types/gen/service/ClickListResponse.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { ServiceClickItem } from './ClickItem.ts'

export type ServiceClickListResponse = {
  /**
   * @type array | undefined
   */
  list?: ServiceClickItem[]
  /**
   * @type integer | undefined
   */
  total?: number
}